+++ csatar-plugins/knowledgerepository/controllers/Locations.php
@@ -3,25 +3,25 @@
 use Backend\Classes\Controller;
 use BackendMenu;
 
-class AccidentRiskLevel extends Controller
+class Locations extends Controller
 {
     public $implement = [
         'Backend\Behaviors\ListController',
         'Backend\Behaviors\FormController',
-        'Backend\Behaviors\ReorderController'
+        'Backend\Behaviors\ReorderController',
     ];
 
     public $listConfig = 'config_list.yaml';
     public $formConfig = 'config_form.yaml';
     public $reorderConfig = 'config_reorder.yaml';
 
-    public $requiredPermissions = [
-        'csatar.manage.data'
-    ];
-
     public function __construct()
     {
         parent::__construct();
-        BackendMenu::setContext('Csatar.KnowledgeRepository', 'main-menu-knowledge-repository-parameters', 'side-menu-item2');
+        BackendMenu::setContext('Csatar.KnowledgeRepository', 'main-menu-knowledge-repository-parameters', 'side-menu-location');
     }
+
+    public $requiredPermissions = [
+        'csatar.manage.data'
+    ];
 }