+++ Changelog.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
-- CS-375 - Created Tool model, controller, backend list and form
-- CS-386 - Created AccidentRiskLevel model, controller, backend list and form
-- CS-385 - Created GameDevelopmentGoal model, controller, backend list and form
+- CS-384 - Created GameType model, controller, seeder, backend list and form
+- CS-371 - Created Location model, controller, seeder, backend list and form
+- CS-369 - Created Duration model, controller, seeder, backend list and form
+- CS-368 - Created Headcount model, controller, seeder, backend list and form
+- CS-375 - Created Tool model, controller, seeder backend list and form
+- CS-386 - Created AccidentRiskLevel model, seeder controller, backend list and form
+- CS-385 - Created GameDevelopmentGoal model, seeder controller, backend list and form
 - CS-385 - Changed migrations, renamed "order" columns to "sort_order"
 - CS-363 - Fixed success/error message issue on back list bulk delete
 - CS-455 - Sort the teams in the menu